backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
langchain
langchain-community
langchain-core
langchain-google-genai
langchain-chroma
chromadb
python-dotenv
python-multipart
pypdf
